+++ packages/tasks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @workglow/tasks
 
+## 0.0.56
+
+### Patch Changes
+
+- Update TaskGraph to add subscriptions for status changes for tasks and dataflows
+- Updated dependencies
+  - @workglow/task-graph@0.0.56
+  - @workglow/util@0.0.56
+  - @workglow/job-queue@0.0.56
+  - @workglow/storage@0.0.56
+
 ## 0.0.55
 
 ### Patch Changes